+++ service/controller/key/machine_deployment.go
@@ -5,8 +5,8 @@ import (
 	"strconv"
 
 	"github.com/dylanmei/iso8601"
-	"github.com/giantswarm/apiextensions/v3/pkg/annotation"
 	infrastructurev1alpha3 "github.com/giantswarm/apiextensions/v3/pkg/apis/infrastructure/v1alpha3"
+	"github.com/giantswarm/k8smetadata/pkg/annotation"
 	"github.com/giantswarm/microerror"
 
 	"github.com/giantswarm/aws-operator/service/controller/resource/tcnp/template"